requirements.txt
fastapi 
uvicorn
dotenv
langchain
langgraph
requests
finnhub-python